feat(io): fix remaining f32/f64 boundaries in pmi_render, hash keys
- pmi_render.rs: to_arr returns [f32;3], cast tol.value + ra_value
- assembly.rs, scene_to_step.rs: [u32]→[u64] hash key types
- mesh_export.rs: PMat4→Mat4 conversion for append_mesh_with_transform,
  verts_f64_to_f32 for STL output, grid_fallback_rate cast

19 remaining errors: cross-crate f32/f64 mismatches

+++ crates/rc3d-io/src/step/assembly.rs
@@ -246,7 +246,7 @@ impl AssemblyContext {
 
     pub fn shell_instances(&self, entities: &EntityIndex) -> ShellInstanceList {
         let mut instances = ShellInstanceList::new();
-        let mut seen: HashSet<(u64, [u32; 16])> = HashSet::new();
+        let mut seen: HashSet<(u64, [u64; 16])> = HashSet::new();
         let mut visiting = HashSet::new();
         for root_pd in self.graph.find_pd_roots() {
             self.graph.collect_shell_instances_dfs(
@@ -453,7 +453,7 @@ impl AssemblyGraph {
         world: AssemblyTransform,
         entities: &EntityIndex,
         out: &mut ShellInstanceList,
-        seen: &mut HashSet<(u64, [u32; 16])>,
+        seen: &mut HashSet<(u64, [u64; 16])>,
         visiting: &mut HashSet<u64>,
     ) {
         // Iterative stack-based DFS — avoids stack overflow on deeply nested assemblies.
